+++ components/forms/fields/EmailValidationForm.tsx
@@ -1,0 +1,163 @@
+'use client';
+
+import { useForm } from 'react-hook-form';
+import { z } from 'zod';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { useEffect } from 'react';
+
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from '@/components/ui/form';
+import { Input } from '@/components/ui/input';
+import { Switch } from '@/components/ui/switch';
+
+import useDragDrop from '../hooks/useDragDrop';
+import { FormElementInstance } from '../interfaces/FormElements';
+
+const propertiesSchema = z.object({
+  label: z.string().min(2).max(50),
+  description: z.string().max(200).optional(),
+  required: z.boolean().default(false),
+  placeHolder: z.string().max(50),
+});
+type propertiesFormSchemaType = z.infer<typeof propertiesSchema>;
+
+export default function EmailValidationForm({
+  elementInstance,
+}: {
+  elementInstance: FormElementInstance;
+}) {
+  const { updateElement, is_quiz } = useDragDrop();
+  const form = useForm<propertiesFormSchemaType>({
+    resolver: zodResolver(propertiesSchema),
+    mode: 'onBlur',
+    defaultValues: {
+      label: elementInstance.question,
+      description: elementInstance.description || '',
+      required: elementInstance.is_required,
+      placeHolder: '',
+    },
+  });
+  useEffect(() => {
+    form.reset({
+      label: elementInstance.question,
+      description: elementInstance.description,
+      required: elementInstance.is_required,
+      placeHolder: '',
+    });
+  }, [elementInstance, form]);
+
+  function applyChanges(values: propertiesFormSchemaType) {
+    const { label, description, required } = values;
+    updateElement(elementInstance.Id, elementInstance.page_number, {
+      ...elementInstance,
+      question: label,
+      description,
+      is_required: required,
+    });
+  }
+
+  return (
+    <Form {...form}>
+      <form
+        onBlur={form.handleSubmit(applyChanges)}
+        onSubmit={(e) => {
+          e.preventDefault();
+        }}
+        className="space-y-3"
+      >
+        <FormField
+          control={form.control}
+          name="label"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Label</FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  onKeyDown={(e) => {
+                    if (e.key === 'Enter') e.currentTarget.blur();
+                  }}
+                />
+              </FormControl>
+              <FormDescription>
+                The label of the field. <br /> It will be displayed above the
+                field
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="placeHolder"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>PlaceHolder</FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  onKeyDown={(e) => {
+                    if (e.key === 'Enter') e.currentTarget.blur();
+                  }}
+                />
+              </FormControl>
+              <FormDescription>The placeholder of the field.</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="description"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Description</FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  onKeyDown={(e) => {
+                    if (e.key === 'Enter') e.currentTarget.blur();
+                  }}
+                />
+              </FormControl>
+              <FormDescription>
+                The helper text of the field. <br />
+                It will be displayed below the field.
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="required"
+          render={({ field }) => (
+            <FormItem className="flex items-center justify-between rounded-lg border p-3 shadow-sm">
+              <div className="space-y-0.5">
+                <FormLabel>Required</FormLabel>
+                <FormDescription>
+                  The helper text of the field. <br />
+                  It will be displayed below the field.
+                </FormDescription>
+              </div>
+              <FormControl>
+                <Switch
+                  checked={field.value}
+                  onCheckedChange={field.onChange}
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+      </form>
+    </Form>
+  );
+}
