+++ package.json
@@ -38,6 +38,7 @@
     "cmdk": "^0.2.1",
     "date-fns": "^3.3.1",
     "framer-motion": "^11.0.5",
+    "json-schema-to-zod": "^2.0.14",
     "lucide-react": "^0.323.0",
     "negotiator": "^0.6.3",
     "next": "14.0.4",
